Draft split of Omniture from clickstream modules
Clickstreams (and more things like this in the future) are event
*generators*. Omniture (and Ophan etc.) are event *consumers*, who subsequently log
the request to the server (or throw it away, or sample every n requests)

As such our code should reflect this.

I've ported Grant's clickstream code to use this pattern.

The Omniture module will listen for clicks etc. and log them using the
Omniture model (the horrible global 's' variable).

The Clickstream module just listens for document click events and
broadcasts a message along with the data-link-name id it's composed.

Neater?