Guardian logo now gets upgraded to an SVG image.
To make any image upgradeable simply put the attribute data-svgsrc="/images/XXX.svg" on it.

Took away the requirement that we are not in base mode for image upgrade. The upgrade is clever enough to know that on an image by image basis (and it is not relevant to SVG)

Also fixed some broken tests