userPrefs now set by internal anchor
1.

I've moved this to use internal anchor, not querystrings for the setting of prefs.

If we start casually sharing some of our experiments I'd rather Google et al. did not see querystrings, which may dilute our rankings.

- http://googlewebmastercentral.blogspot.co.uk/2007/09/google-duplicate-content-caused-by-url.html
- http://www.webpronews.com/matt-cutts-on-penguin-and-internal-links-2013-04

2.

In writing the tests I've fixed the bug that means externally set variables are now stored as the correct data type (string, int, bool).

At the moment they are all stored as strings.