Added s-maxage, remove public
i. Our CDNs will use 's-maxage' to cache the resources.

ii. The clients will use the 'max-age' value.

iii. When we switch to Varnish (on a single domain) we either need to remove 's-maxage' on outbound requests or add the 'private' declaration.

In this regard I thinking the 'public' declaration is spurious.