Proof of concept for auth ActionBuilder
This allows Identity to easily create actions that require a logged-in
user.

If the user is logged in, the details extracted from the cookie are
included on a User object added to the request parameter as
request.user.

If the user is not logged in, we issue a redirect to the dentity
signin page.
TODO: The signin page's return URL won't be right at the moment.