Switches now have an sell-by-date
The idea of an automated expiry date has worked well for our AB testing. It forces people
to think how long a feature should last in production, rather than the indefinate default.

If a feature has expired two things happen.

i. The feature is disabled.
ii. The tests will break, at which point someone should remove it or seek justification as to why it should live on.

The sell-by-date is optional at the moment (it default to 'the future'), but we'll make it permanent once the concept is understood by everyone.