Effective Measure tag for Australians
"The second [tag] is for Effective Measure which provides an opportunity to obtain audience and campaign data for our Australian audience which we
desperately need in the absence of other data at this point in time!  Ian Gibb (Commercial Insight) has already initiated discussions with both the
provider and our Producer (Madhvi Pankhania) and we are now awaiting next steps."

3 month trail. Contact Matt Kopka & Sam Yee.

We had a 3 week delay on this while Dave Boxall investigated the effectivemeasure.net host name from ScanSafe. He approved it.