Use the more extensive User information from the GU_U cookie
Instead of providing user-data from the very minimal SC_GU_U cookie - which
is a 'secure' cookie and thus less susceptible to eavesdropping - use the
more extensive GU_U cookie, once it's been checked that the user-id
matches on both cookies. This makes more user-data available to us without
having to make an Identity API call.

Both cookies are signed, so they can't be faked, and SC_GU_U mitigates the
danger of an attacker using a stolen cookie.