Make the second nav item a 'child' of the first.
Google does not seem to be picking up our breadcrumbs...

I have done some poking around the dark corners of the internet and I think it is because our nav is two tiered instead of a simple list.

I have added the 'child' markup so that google now sees the nav slightly differently.