Clean invalid XML characters from CAPI fields
When producing RSS a few of the CAPI contain invalid XML characters, so we need to clean them.

This is currently the sole cause of 500 errors on NGW RSS feeds.

From testing a few hundred feeds the problem (?) seems localised to only a few fields, so I've only fixed them.