Start of fixing IDN-1736: 'Resend validation email' for logged-out users
We've had complaints that the 'Resend my verification email' button does
nothing. This has been true if you are NOT signed in:

https://profile.theguardian.com/verify-email/useThisBrokenToken

...the button says resend email, but doesn't actually know who you are!
So it can't send the mail.

This new page will require the user to be logged in, and we will direct
verification-failing users at it if they're not signed in:

https://profile.theguardian.com/verify-email

Once logged in, it will immediately send a fresh verification email to the
associated account.

https://jira.gutools.co.uk/browse/IDN-1736