Proper player cards
Player cards can be created for any player (not just hard-coded to the
premier league). Additionally, the stats are specific to the selected
team and competition instead of being for that player generally. This
also means that dates are no longer hard-coded for the stats since
they use the selected cmopetition's start date.

Also pulled out the FutureZippers, certainly one of my sillier ideas
ever.