Add support for uploading Avatars [#IDN-1774]
Allows users to upload new Avatars for their profile:

https://profile.thegulocal.com/public/edit

The avatars are sent for moderation in the Discussion tool, so they're not
available immediately.

The Avatar upload system we're hooking into /could/ be described as
legacy. The fact that it uses GAE is not the problem, the dependency on
the Guardian DocRoot (rather than S3) is closer to the mark.

https://jira.gutools.co.uk/browse/IDN-1774
https://sites.google.com/a/guardian.co.uk/discussion-project/
https://sites.google.com/a/guardian.co.uk/discussion-project/avatars/avatar-upload