Add player cards by date
Previously player cards were tied to a competition. This lets player
cards be created by instead specifying a start date. The stats will be
for all matches by the given player on the given team. This is handy
for the world cup because otherwise there'd be no stats at the start
of the competition.