Fixed ad units
We used to send shortest possible adunit strings to DFP. So an article
with the “lifeandstyle/live-better” tag would just send the
“lifeandstyle” adunit.

This created confusion as we tried to figure out what to leave out.

Since DFP allows our adUnits to nest, sending the full list like
“lifeandstyle/live-better” really doesn’t hurt us, and it makes the
code much easier to read.

Well, except for that bit with the editionalised fronts.