Fix for jumping trails
When an article is dragged over parts of another article that have their own context (like the boolean states) a request is sent to add the new article at the bottom of the parent group instead of above the article.

There's also some refactoring to split the view/action from the drop/position logic.