Added "simple" navigation metrics.
The intent is to have really quick feedback for when we start lazy loading some components on the bottom of the page.

(this also deletes some live blog metrics I put in a while back and no longer use).

by using a sample of browsers (those that support sendBeacon) this will give us a near real-time view of...

1. Percentage of people going to view another Guardian page.

2. Percentage of people interacting with the page in a way that does not cause another page to load (e.g. clicking a tab).

 3. Percentage of people clicking "Related content"

 4. percentage of people clicking "more form this series"