Sparklines
First simple implementation of sparklines.

Data is requested from ophan in batches, after the front is fully loaded.

For now, only the articles inside a front will displaoy sparklines and the data is filtered by referrer-url (the front).

This is a first step, we need to verify that both ophan and the tool can handle the extra load.

Everything is behind a switch (same switch as before) but it's possible to override that switch adding the url parameter `?sparklines=please`