Fix CSS source maps
CSS source maps are currently broken. This was happening because `grunt-px-to-rrem` removes the source map annotation (when PostCSS rebuilds the CSS AST). This can be fixed by upgdating the transitive dependencies. I tried, but it's very fiddly and made even more difficult by the shortsightedness in which the grunt task and its dependencies were authored (as per the entire Node ecosystem, it seems).

As a quick fix, we can just prevent the task from running in dev—when source maps are in use.

Can I ask why we added the converter in the first place? /cc @phamann

Fixes #8508