Re-word social sign-in confirmation message
When using social sign-in, a banner at the bottom of article pages informs
users that they're logged in with $socialprovider. The confirmation text
reads:

"Welcome $name, you're signed into the Guardian using Facebook, or sign
out".

That seemed a bit odd in terms of style, so this commit updates it to:

"Welcome $name, you're signed in to the Guardian using Facebook. Sign out."

The sub-clause in the message didn't sound right.