Re-auth expired sessions
When sending an API request, if the token is expired, panda sends back a 419 error.
In that case, try to silently re-auth using panda-session and retry the previous request.

If the sencond attempt doesn't work, redirect to login page.

Lots of file changed because authed-ajax is now using promises instead of $.Deferred. I wrote a unit test for everything that changed, I hope I haven't missed something.