Prevent overlapping form submissions
We are currently seeing quite often users upload multiple avatars within
a short time period (< 1 second), probably because they are clicking the
upload button multiple times. The previous behaviour only disabled the
form once the upload had completed, which was too slow.

This commit disabled the form immediately after the user clicks the
button to prevent overlapping submissions. We re-enable the form if the
upload returns as an error.