Some environments don't support or don't always expose the console object
I believe this will fix error https://app.getsentry.com/the-guardian/client-side-prod/group/69062228/

I found the issue whilst trying to understand how we broke tracking in IE9. I was able to find the issue in Sentry by setting "browser" to "IE 9.0" and "featureName" to "common": https://app.getsentry.com/the-guardian/client-side-prod/?status=&feature=common&browser=IE+9.0. It's the third most frequent error for that filter search. (Although, beware that Sentry defines frequency as an aggregate, not per filter search.)

It was introduced by 4d0e1395f40990b3e946813efef4f675966645e9, in PR #9485.

The reason for this error is because, [whilst `window.console` is supported in IE8 and 9, it is not defined unless you open the developer tools](http://stackoverflow.com/a/5473193/1082754). FML.

/cc @gklopper @janua