Send beacon for unique visitors in JspmTest and JspmControl
We want to know if the ~1.5% drop in unique visitors and 3% drop in page views for JspmTest is because of an error in SystemJS or something beforehand (i.e. inaccurate bucketing or bounce due to bigger payload). We can use the data from these beacons and compare it with the data in Omniture.

/cc @gklopper @sndrs