Crosswords: remove auto focus
We shouldn't be stealing focus when the page loads, because keyboard/screenreader users will have to navigate back up to the top of the page.

There is also a bug whereby, if we autofocus 1-across and the user taps the first cell to open the keyboard, the selection alternates to 1-down. This will be avoided if we don't autofocus.