Facia tool switches about to expire
FaciaDynamoArchive is not used anymore, the new facia-tool does not even have access to that DB. We'll figure out another way to do auditing. Might be a good idea to delete the old database.

FaciaToolPutPrivate was never turned on, we'll switch it on by default when we migrate collection to a new bucket, for the moment we can live with public collections