Fix adverts on live blogs
This was a problem with one of the policies in commercial-feature-policies. The 'non article pages' policy was triggered by a page which had a contentType other than 'Article' *and* was not a live blog.

This change classifies a non-article page as one without an 'Article' content type or one that is a live blog.