Fix clue URL bug
When selecting a cell within a new clue (a clue that is not currently foccussed), the clue URL code was causing the selection to be reset to the first cell within that clue.

We update the URL when you focus a new clue, but in the event listener for `hashchange`, we only want to update the focus if the clue has not already been foccussed (i.e. by selection). An example of when we do want to update the focus is when you navigate using the browser history.