Prevent app banners from appearing unstyled and FOUC
Fixes #10910

This is a race condition: if the CSS loads before the JS executes, you don't get a flash, otherwise you do.

One solution would be to add this CSS to the inline CSS, but that would increase the inefficiency and bloat of our inline CSS, as we hardly ever show these banners.

This solution will force the JS to wait until the CSS has loaded before injecting the DOM element.

I wonder if we could move more CSS out of inline and towards this model of waiting to execute JS.