Remove self-help section from tech feedback page to better gauge crashing
We haven't had any feedback about crashing since the end of September. Is this because users are now guided to the self-help page? Feedback is a good indicator of whether we still have a problem, so this temporarily takes away the self-help section to help us better gauge whether users are still having problems.