Header-bid adverts on US edition of www.theguardian.co.uk
If you read the US edition of www.theguardian.com on desktop, you will have a 10% chance of seeing adverts selected by 'header bidding'. This means that the browser will ask third parties what they can pay for an advert, send the details of the best bid up to DFP and then await its decision whether to award the adslot.