video embeds obey advertising rules of parent video
Before:
Video is published with pre-roll=true
Video is embedded
Video pre-roll is then changed to false
Embed still has pre-roll=true

(and vice-versa for video published with pre-roll=false)

After:
I've added a field to /info.json which is used to determine if advertising should be blocked on the video. Therefore, similar to how expiry is calculated, the embed follows the parent video.