Allow the Member's Area redirect to apply on all parts of the site
Previously, only articles that had '/membership/' in the url - like this
one:

http://www.theguardian.com/membership/2015/apr/27/the-spin-london-cycle-festival-save-mop

...would actually have the Member's Area restriction take effect - the
setting wouldn't affect articles under theguardian.com/film, for instance.

We're planning to use the Member's Area on a wider basis around the site,
and also apply it to the Mobile Apps premium content, so this '/membership/'
restriction (initially in place to reduce possible editorial accidents)
needs to go.

See this doc for some more information on Member's Area content:

https://docs.google.com/a/guardian.co.uk/document/d/1oQhe3PwhcJdmmm_q6_9issBqOA69vE0hMcT28WztQsA/edit?usp=sharing