Remove server-side A/B tests variants
We used to have 10 variants (each of the representing 1% of our traffic)
which you can assign server-side a/b tests to.
However this was creating 10 CDN 'buckets' for each url and therefore
reduced our cache/hit ratio and cost us a lot of money.
We are moving away from this system: this patch is removing the
remaining variants and refactor slightly the MultiVariateTesting file