always show video over sublinks
Currently, if a card is showing video and has two or more sublinks, the video is hidden (for reasons unknown).
This is different for images - cards with an image and two sublinks shows all elements.

This change brings video and images inline - videos are always shown regardless of the number of sublinks.