Track clicks using GA as well as Omniture
Adds support for tracking the following events using GA:

* in-page link clicks
* internal link clicks
* external link clicks
* 'interactions' (e.g. printing the page)

Also refactored the existing code (which was heavily coupled to
Omniture), introducing a click-tracking layer to abstract over Omniture
and GA.

The 3 types of click events are handled as follows.

In-page links are easy: just fire off an Omniture/GA event

Internal links: write the necessary data to session storage, send a
click event on next page load. This is done in a slightly clunky way:
first GA loads, finds the data in session storage, sends an event. Then
Omniture loads, also finds the data in storage, sends an event, then
deletes the data from storage.

External links: fire off Omniture and GA events, and rely on Omniture's
magic to delay navigation for 500ms to give the events time to complete.
Not sure exactly how Omniture does this, but when we remove Omniture
we'll need to re-implement this behaviour.