Fix Google Analytics in video embeds
Embeds were throwing a JS error and failing to send the GA hit because
they relied on some Ophan-related config code that was not being run.

Note: with this change, the Ophan pageview ID is now generated by
frontend and passed to the Ophan JS via config, just like on normal
frontend pages.