messaging for geo-blocked video
We can currently geo-block videos, this adds some messaging to the video element if you're in a blocked region.

The videos are blocked at the CDN level. Annoyingly, videojs doesn't make it easy to view the status code, so lets just make a HEAD request to get the response code. If its a 403, then we say the video isn't available. Any other status code is handled how they usually are.