Set fixed date for `now` game date in MatchTestData
tl;dr: Don't make tests rely on current datetime. :)

One FixturesList feature consists in displaying a list of game matching
some criteria. The list is restricted to the eligible games scheduled
within 3 days of the earliest game.
In MatchTestData, mocked games are created with a relative date base on
`now`.
The earliest game date used to be now+50mins
Therefore if running the tests after 23:10, the earliest game would have
a date set to 'tomorrow'. FixturesList will list games for 3 days
starting tomorrow and not today. Tests fails then since the games in the
list are not the one expected.
Example:
It is Aug 4th at 23:30. First game is set to Aug 5th at 00:20.
FixturesList then returned all the games schedule on Aug 5th, 6th and
7th. However the test is expecting games from Aug 4th (now) to Aug 6th

The simple fix is to set a fixed date for `val now` so tests don't rely
on current datetime and are perfectly repeatable