Make Membership engagement banner into an A/B test with Contributions
For 50% of the audience, show the Contributions message, rather than
Membership.

https://trello.com/c/HlD5H68N/325-engagement-banner-a-b-test

A fair bit of refactoring of the existing Membership engagement
banner functionality was necessary, to make it fit into the
AB testing framework - which in particualar has a synchronous
`canRun` method, which is problematic given that the banner
is not supposed to display at all if the asynchronous check
for an adblocker comes back positive.