Allow the engagement banner to appear on all pages, not just Articles
The engagement banner was restricted to Article pages as far back
as it's origins in https://github.com/guardian/frontend/pull/9367
in June 2015, but having checked with Charles Minty, Paul Brown
and Julia Bellis, we can't find any solid reason why.

Everyone is happy to try enabling it! The banner appears to work
fine on fronts, and does not obscure full-screen video on video
pages like http://localhost:9000/world/video/2016/jul/21/nice-attack-planned-for-months-involved-accomplices-says-prosecutor-video