Replace unnecessary JSON conversions and JsonComponent constructors.
There is now a generic JsonComponent for getting a 200 JSON response
given a JSON-serializable object.

There is probably more cleaning up that can be done. Given that there
is an HTML specific JsonComponent constructor, a lot of the calls that
create a `("html -> html)` input can be simplified, but only in the
cases where `html` is an `html` object, thus triggering said specific
constructor.