Removes unsafe assumption that a Json serialized A can become JsObject.
Previously, some objects were passed in explicitly as JsObject. The
apply method that dealt with these then added on the RefreshStatus.

This behaviour needs to be retained, but given a type A with a JSON
Writes method, we cannot assume that we will be able to convert to
a JsObject and add on the RefreshStatus e.g. JsString (a sibling of
JsObject).

This changes introduces a pattern match to add the RefreshStatus only
when the A passed in is JsObject.