Only add a badge to the first container that ISN'T a thrasher on fronts.
Previously we add a branded badge to the first container on a branded
front. However, we can have thrashers at the top of such a front, and
these should not have the badge placed in them. We can have multiple
thrashers on a single front and so the most robust way to tweak the
current behaviour is to persist the knowledge of whether each
container is a thrasher or not and then, when rendering each container,
pass in a property specifying whether a given front should be badged.