updated according to comments
Have implemented almost all comments.
Exception is AMP testing as a seperate section.
While the reasoning is valid, I really want to highlight the importance of testing on AMP for the moment as it 
is impacted by more things than you might expect. As a result, on second reading I'm not only keeping it in, I'm making it more prominent in the flow.