Treat 404 from members data API as valid
Before the refactoring to use fetch-json, any response was treated as valid, including the 404 that members data API uses to indicate that the user is not a known member. This reintroduces that behaviour, but limits it to just 404s, so 5xx statuses don't result in us setting a cookie indicating the user is not a paying member.