Add module to check an A/B test against the targeting tool
The contributions team is planning to use the [targeting
tool](https://github.com/guardian/targeting) to manage the tags used in
Epic messages, so we can control where it shows up without having to
make any code changes. To help with this I've added a small module that
checks the campaigns in `guardian.config` (populated by the targeting
tool) against a given A/B test, which will give us a value to use in the
test's `canRun` criteria.