Remove useless 'secure' conf param ans isSSL js page config
isSSL was only used in one place to historically prevent ophan to be
initialized on identity page (which were on https). I guess ophan was
not on https at the moment.
This check is useless now ophan is on https (as well as all pages of the
guardian). Hence removing it.
'secure' configuration parameter was only used to set isSSL. Therefore
removing it now it is not used anywhere