Refactor the engagement banner and it's A/B tests
Autogen the ABTest switches for the engagement banner

Add AB tests back into engragement-banner

Autogen the ABTest switches for the engagement banner
Set default campaign codes for non A/B test banners

--
We're using lodash `find`, as the `Array.prototype.find()` has patchy
support:

https://lodash.com/docs/4.17.2#find
https://developer.mozilla.org/en/docs/Web/JavaScript/Reference/Global_Objects/Array/find
mbership-engagement-international-experiment.js
--

It was a bit of a surprise to me that Contributions engagement banner
currently does rotating colours- it used to be a permanent yellow, I think-
but the Contributions team don't seem to mind. It is a bit simpler
for it to match the Membership banner rotating-colour strategy.