Add Banner test 13, with message rotation and 95p variants
This PR introduces a new convention that the `messageText` param can be
either a simple string or an array of strings. If the `messageText` param
is an array, the messages supplied will be rotated through on successive
page views (much like the colours already are).

As well as that, we are testing a variant that is identical to the control,
but specifies 95p a week (ie mentioning a smaller weekly figure first,
though payment is monthly or annual), which research from other media orgs
indicates might be a good way to draw people in.


https://trello.com/c/XNDFl8SZ/215-engagement-banner-test-13-test-3-rotating-messages-vs-1-single-message-95p-price-variant