Moving breaking-news to onward
I made a lot of mistakes when creating the breaking-news feature, from the
internal design choices (using AkkaActor for instance) to the location of
the endpoint (in admin-jobs)
Ideally we would need to refactor the breaking-news feature but this
patch is only addressing the location issue.
Breaking-news has nothing to do in admin-jobs. This patch is moving it
in onward, even though it might not be the best place for the POST /news-alers/alert endpoint, at least it is logical to have the GET /news-alert endpoint in onward (and because of the bad design choice we cannot simply break the 2 endpoints into 2 different apps)