Add Acquisitions test selector utility
This tool prevents multiple acquisitions (formerly contributions) tests
clashing with each other by working out which one should be active. It
is not enough for us to rely on the audience segments since we might
have multiple tests running on the same segments, but targeted at
different tags via the targeting tool.