Make sure email forms appear correctly for Sleeve Notes sign up (#15979)
We are going to run a test where new subscribers to sleeve notes get either the new or old version of the email, and so we need two new ids for old/new versions of the
email for these new subscribers. And these new ids need to have a 'review' tone so
that the sign up form appears correctly in the sign up page.