fix the poster image not hiding when video is playing in Chrome
For some reason, Chrome 57 doesn't like the current css, I think its to do with the background-size property but can't confirm.

As a really hacky fix, lets make the height of the poster 0 to "hide" it when the video is playing/paused.

Removing background-size wasn't an option because the poster image needs to stretch to fill its container (background-size: cover).

Using opacity also wasn't an option as this effects child elements and made the "Watch now" button transparent.