LinkTo: Making editionalisation of a link optional
Section title returned by onward should not be editionalize because
onward requests are not cached per edition (no cookie is passed to
api.nextgen.guardianapps.com). Therefore all clients will get the same
content.
Not editionalizing the title link means the editionalization would be
taken care of by a redirect when the link is clicked
This patch is adding a way to explicitely tell `LinkTo` not to
editionalise a link.