Make epic in liveblog pass AMP validation (#16629)
Fixes an AMP validation issue in #16519.

Items in an amp-live-list are required to have an id and data-sort-time.
https://www.ampproject.org/docs/reference/components/amp-live-list#items

I was inserting a placeholder element for the epic which didn't have these. So I've opted to ditch the placeholder element and instead add a class to the block itself, then insert the epic after that selector.