Initialize more than one rendering actor
The current number of account (3) has been choosen quite arbitrarely.
We will need to benchmark what is the best value depending on the
performance of the rendering. We might also want to set a different
count based on the kind of requests (ie: different app could have
different amount of rendering actors)