Unregistering service-workers when hitting preview login page
Google auth redirection in preview is handled by writing in a cookie the last
requested urls. If a service worker is installed the browser might try
to refresh it and therefore save in the cookie the url of the
service-worker itself. When the user presses login they are then
redirected to service-worker.js :(
This patch ensured no service worker is registered when the user is not
logged in preview.