membership-engagement-banner-tests to ES6 (#17624)
Because of the stricter linting for ES6 code, I'm not allowed to leave the unused `EditionTest` class. Which is fair enough, because having it there as a "template" for other people to copy & paste is pretty weird. It should be made into a `MembershipBannerTest` class which people can extend or instantiate, or turned into a factory method like `makeABTest()` in `contributionsUtilities`.

But for the time being I've just deleted it. Whoever creates the next banner test can decide on the best approach. I've left the rest of the scaffolding in place

### Also in this PR:
- `epic` property removed from `ContributionsABTest` type and type renamed to `EpicABTest`
- `AcquisitionsABTest` type added to cover tests which might be engagement banner or epic tests
- `isEngagementBannerTest` property removed from `ABTest` type