Use $ReadOnlyArray  type for immutable covariant arrays (#17640)
Convert some `Array<T>` (aka `T[]`) types to `$ReadOnlyArray<T>`

This has the following advantages:
- Enforces immutability of these arrays, which reduces the likelihood of bugs
- Allows arrays to be covariant (thus fixing an issue in #17637 where I had to use `Object[]` to get around the fact that mutable arrays are not covariant)

The latter advantage only really applies where the parameterised type `T` is a subtype or supertype of another since otherwise the variance of higher-order types using `T` is irrelevant. So it's relevant for `ABTest` since we have subtypes `AcquisitionsABTest` and `EpicABTest`. It's less important for, say, `OphanProduct`