always show play button on fronts
Previously, we were not adding the play button to small cards when they had a media atom as main media. This is different from old style videos. In old style videos, videos would play inline on fronts. Although Editorial accept that we cannot play YouTube videos inline, they still want the play button to show; so much so infact that its been known to burn it into the image!