Show more button should always be visible on football features/results
With or without javascript the button should always be displayed.
The implementation in this commit is a workaround only modifying the
match page template and css, a correct way to achieve the same results
would be to change how all the 'more button' work and explicitely
specify it should be hidden when js is disabled.
The correct fix will be implemented in a subsequent pull request.