Add linter for disallowed strings (#17980)
Adds a build step called `check-for-disallowed-strings` which searches the codebase for disallowed strings based on the rules in `disallowed-strings.js`. Each rule has a format like this:

```
{
    regex: /membership\.theguardian\.com/,
    message: 'Raw URLs to the membership site are not allowed, use getMembershipURL instead',
    maxOccurrences: 1,
    pathspecs: ['*.js', '*.scala', '*.html']
}
```

`pathspecs` are [Git pathspecs](https://git-scm.com/docs/gitglossary#gitglossary-aiddefpathspecapathspec).

`maxOccurrences` would in most cases be 1 to avoid duplicating hard-coded values. But it could be 0 to disallow some string entirely. It could also be greater than 1 if you allow the value in a few places (e.g. clientside & serverside).