Install the click event listener on each question rather than on the container
Before change in previous commit, the event listener was installed on each
question, and the code handling the response was expecting this, as it using the event current target instead of the event target.
 
Fixing this issue led me reading bean source code, which I wish I would not have done. There is no words I know in english to describe how unexpected and difficult to read its source code is, due to everything being dynamic..