remove Missing Video Encodings job
There is an akka job that runs every so often that gets the recently published video pages and then ensures the urls resolve. If they don't an email, via an SNS topic, is sent.

Media Atoms, primarily, use a YouTube embed and so this job is no longer needed.

The dynamo table `missingVideoEncodings` and the SNS topic `arn:aws:sns:eu-west-1:642631414762:frontend-missingVideoEncodingsNotificationTopic` don't appear to be cloudformed so would need to be deleted manually.