Prevent most popular from overlapping most viewed on short articles.
The nature of the sticky mpu remaining in the viewport for a period of time means that most popular gets pushed further down in the right hand slot.
On short articles this causes most popular to get pushed behind most recent. Therefore for short articles I have set a threshold with which the most popular container will only be shown if said threshold is exceeded.