Add World Cup 2018
Includes ordering to build wallchart in spider form.

The orderings have to be right so all the stages line up on the
"spider" wallchart. This is provisionally correct, but assumes the
Semi-Finals (we'll need to check this with PA before merging).

TimeZones are required on the match dates - we use UK times for all
the football stuff.