theme YouTube player based on pillar
It is now possible to set the colour of the control bar in the YouTube player! Set the value based on the content pillar.

Note: we can't change the colour of the text (it is always white), therefore we can't use the special report yellow as the contrast isn't high enough.