Improve KnockoutSpider logic
... in preparation for adding the World Cup 2018

If orderings are provided but do not apply to the matches in a round,
that round will fall back to a knockout list.

This is not technically necessary because orderings are manually added
(and presumably whoever adds them can check it works), but it means
the application fails sensibly, rather than displaying totally broken
markup. There are two cases where this happens:

* while developing, the person that adds the orderings does it wrong
  (e.g. forgets the timezone). The totally broken display was super
  confusing, made it hard to work out what the actual problem was.
* The KnockoutSpider requires league table data, which is a separate
  endpoint in the PA data. Sometimes, PA do not enable that endpoint
  for a competition. This means the fallback in those cases (or if the
  leaguetable ever fails to load for whatever reason) makes some
  sense.

Also involves a minor refactor for clarity, as well as a new test case
for this behaviour.