Add RH col with ad to crossword at wide breakpoint
If ads are enabled we add a secondary-column to the crossword page
that only appears at the largest breakpoint. It contains an ad slot
that will accommodate ads up to 300px wide.

The crossword layout has been tweaked to fit the ad in at this
breakpoint, based on the design in the ticket.