Fix height of facia football snaps
There is JS, in football-snaps.js, that will adjust the height correctly but
this was being overridden by the 'important' styling. In addition, flex
styling is needed to ensure the height grows/shrinks correctly. This is
present at non-mobile breakpoints but is needed at mobile too because the
'more' link requires it to be positioned correctly in the typical case where
the table is larger than container.