Lazy load images on fronts
This change significantly improves initial page weight on
fronts (perhaps up to 1mb or more), which significantly
improves time to interactive on mobile devices/those with a
poor connection.

Images in the top two containers are eagerly loaded to ensure
images in the initial viewport load as quickly as possible.

The feature is behind a 0% AB test initially to enable fuller
testing on PROD before rolling out to a larger audience.

FAQ:

Q. Will images 'flash' when scrolling?
A. Sometimes yes. Images will start loading *before* hitting
the viewport but depending on connection/scrolling speed, they
may not have loaded by the time they come into view. They will
then 'flash' into place as they load. It is arguable that
users are used to this behaviour (many other sites do it) but
I'm not a UX expert! To improve the UX here, we fade in images
to make it seem smoother/more deliberate.

Q. What happens for users with Javascript disabled?
A. They get non-lazy behaviour, implemented using noscript
tags.

Q. What happens if Javascript errors - will images fail to
load?
A. It is definitely possible that images fail to load below
the viewport in this case. However, lazy loading is triggered
by events, which should progress even if other Javascript
errors. The exception is if other JS is slow/blocks the main
thread.

Q. Why not apply this to articles/galleries too?
A. For regular articles, which typically have only a couple of
images, the benefits are less significant. But, more generally,
we need to think more here, as there are many types of articles.

Q. Does this significantly increase the JS bundle size?
A. No. We already include the necessary polyfills, and the
library we use (lozad.js) is < 1kb and could be inlined if
necessary.

Q. Aren't browsers about to implement this behaviour natively?
A. Yes, at least Chrome is. However, it's not live yet, and
many of our users rely on browsers that are not implementing this
in the immediate future. It would be great to rely on native
behaviour here though when it becomes possible. There is a good
writeup of the Chrome proposal here:

https://docs.google.com/document/d/1e8ZbVyUwgIkQMvJma3kKUDg8UUkLRRdANStqKuOIvHg/edit#