Prevent email generation failure (#20848)
## What does this change?

We have problem by which, in some circumstances, emails are generated without their embedded styles. When this happens, an endpoint like this one

https://www.theguardian.com/politics/2018/jul/16/theresa-may-ready-cave-in-hardline-brexiters-customs-demands/email

Present an HTML document without the embedded styles

![screenshot 2018-12-14 at 11 51 29-cut](https://user-images.githubusercontent.com/6035518/50002441-0b083800-ff98-11e8-8f4c-7a5d5e4ba494.png)

Investigation revealed that the problem was caused by a CSS Parser that is not thread safe. 

When that library was first used, we knew there was a problem ( https://github.com/guardian/frontend/pull/15720 ), but it was not handled the best way and a `Retry` was used to mitigate it. Today's refactoring prevent the same object from being reused between requests.

It is not impossible that a follow up PR will come to get rid of the Retry all together (more investigation is required to make sure it is not needed). 

### Tested

- [x] Locally
- [ ] On CODE (optional)